+++ lib/chef/provisioning/driver_init/fog.rb
@@ -1,0 +1,3 @@
+require 'chef/provisioning/fog_driver/driver'
+
+Chef::Provisioning.register_driver_class("fog", Chef::Provisioning::FogDriver::Driver)