+++ lib/chef_metal/driver_init/fog.rb
@@ -1,3 +1,0 @@
-require 'chef_metal_fog/fog_driver'
-
-ChefMetal.register_driver_class("fog", ChefMetalFog::FogDriver)