+++ CHANGELOG.md
@@ -1,5 +1,29 @@
 # Changelog
 
+## 0.13.1 (3/4/2015
+
+- Fixed floating assignment for Openstack (@tyler-ball)
+
+## 0.13 (2/25/2015)
+
+- Support for ~/.aws/credentials (@miguelcnf)
+- Crash fixes for mixed machine options
+
+## 0.12 (1/27/2015)
+
+- Allow user to pass private key data in machine_options[:ssh_options][:key_data] (@squaresurf)
+- [AWS] Add IAM role support for AWS via fog:AWS:IAM (@justindossey)
+- [Rackspace] Prevent bootstrap options from being destructively modified (@kanerogers)
+
+## 0.11 (11/5/2014)
+
+- WinRM support for AWS!
+- Work with Chef 12
+
+## 0.10 (10/29/2014)
+
+- Rename to chef-provisioning-fog
+
 ## 0.9 (9/5/2014)
 
 - [AWS] Per-region default AMIs!